Moved tests to test_axelrod_first.py, removed test_graaskamp.py
